+++ docs/Whats-New.md
@@ -631,6 +631,7 @@ HideShakeEffects=false           ; boolean
 - [Show game time](User-Interface.md#show-game-time) (by Trsdy & Ollerus)
 - Provided a toggle for whether the landing direction in default scenarios does not use the building direction but follows `[AudioVisual] -> PoseDir` (by Noble_Fish)
 - [Separate the definitions of default direction for aircraft production and landing in the field](Fixed-or-Improved-Logics.md#separate-the-definitions-of-default-direction-for-aircraft-production-and-landing-in-the-field) (by Noble_Fish)
+- Change target Owner on warhead impact (by Fryone)
 
 #### Vanilla fixes:
 - Fixed sidebar not updating queued unit numbers when adding or removing units when the production is on hold (by CrimRecya)